+++ example/bin/injector.dart
@@ -3,7 +3,8 @@ import 'package:injectable/injectable.dart';
 
 import 'injector.iconfig.dart';
 
+
 final getIt = GetIt.instance;
 
-@injectableInit
-void configureDependecies() => $initGetIt(getIt);
+@InjectableInit(generateForDir: ['lib'])
+void configureDependencies() => $initGetIt(getIt, environment: Environment.dev);