+++ example/bin/injector.iconfig.dart
@@ -1,0 +1,12 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// InjectableConfigGenerator
+// **************************************************************************
+
+import 'package:example/lib_service.dart';
+import 'package:get_it/get_it.dart';
+
+void $initGetIt(GetIt g, {String environment}) {
+  g.registerFactory<LibService>(() => LibService());
+}