+++ example/lib/lib_service.dart
@@ -1,0 +1,4 @@
+import 'package:injectable/injectable.dart';
+
+@injectable
+class LibService {}