+++ example/test/test_service.dart
@@ -1,0 +1,6 @@
+import 'package:example/lib_service.dart';
+import 'package:injectable/injectable.dart';
+import 'package:mockito/mockito.dart';
+
+@injectable
+class TestService extends Mock implements LibService {}