+++ injectable/lib/src/injectable_annotations.dart
@@ -1,12 +1,23 @@
 class InjectableInit {
-  const InjectableInit._();
+  // only files exist in provided directories will be processed
+  final List<String> generateForDir;
+
+  const InjectableInit({this.generateForDir = const ['lib', 'bin']})
+      : assert(generateForDir != null);
 }
 
-const injectableInit = const InjectableInit._();
+const injectableInit = const InjectableInit();
 
 class Injectable {
+  // The type to bind your implementation to,
+  // typically an abstract class which is implemented by the
+  // annotated class.
   final Type as;
+
+  // an alternative way to pass env keys instead
+  // of annotating the element with @Environment
   final String env;
+
   const Injectable({this.as, this.env});
 }
 
@@ -26,6 +37,8 @@ class Singleton extends Injectable {
   }) : super(as: as, env: env);
 }
 
+const singleton = const Singleton();
+
 /// Classes annotated with @LazySingleton
 /// will generate registerLazySingleton func
 class LazySingleton extends Injectable {
@@ -35,7 +48,6 @@ class LazySingleton extends Injectable {
   }) : super(as: as, env: env);
 }
 
-const singleton = const Singleton();
 const lazySingleton = const LazySingleton();
 
 // Used to annotate a constructor dependency
@@ -50,13 +62,14 @@ class Named {
   // you can pass a type and it's name will be extracted
   // in during generation
   final Type type;
+
   const Named.from(this.type) : name = null;
 }
 
 const named = const Named("");
 
-// used to annotate dependecies which are
-// registered under different envionment
+// used to annotate dependencies which are
+// registered under different environment
 class Environment {
   final String name;
 
@@ -71,6 +84,7 @@ const dev = const Environment(Environment.dev);
 const prod = const Environment(Environment.prod);
 const test = const Environment(Environment.test);
 
+@Deprecated('Use @Injectable(as:...) instead')
 class RegisterAs {
   // The type to bind your implementation to,
   // typically an abstract class which is implemented by the
@@ -78,7 +92,7 @@ class RegisterAs {
   final Type abstractType;
 
   // an alternative way to pass env keys instead
-  // of annotated the element with @Environment
+  // of annotating the element with @Environment
   final String env;
 
   const RegisterAs(this.abstractType, {this.env});
@@ -94,7 +108,7 @@ class FactoryMethod {
 const factoryMethod = const FactoryMethod._();
 
 /// Marks a constructor param as
-/// facotryParam so it can be passed
+/// factoryParam so it can be passed
 /// to the resolver function
 class FactoryParam {
   const FactoryParam._();