+++ injectable_generator/lib/generator/lazy_factory_generator.dart
@@ -1,7 +1,14 @@
 import 'package:injectable_generator/dependency_config.dart';
-import 'package:injectable_generator/generator/factory_generator.dart';
+import 'package:injectable_generator/generator/register_func_generator.dart';
 
-class LazySingletonGenerator extends FactoryGenerator {
+class LazyFactoryGenerator extends RegisterFuncGenerator {
+  final isLazySingleton;
+  final String funcName;
+
+  LazyFactoryGenerator({this.isLazySingleton = false})
+      : funcName =
+  isLazySingleton ? 'registerLazySingleton' : 'registerFactory';
+
   @override
   String generate(DependencyConfig dep) {
     final initializer = generateInitializer(dep);
@@ -12,7 +19,9 @@ class LazySingletonGenerator extends FactoryGenerator {
     }
 
     final asyncStr = dep.isAsync && !dep.preResolve ? 'Async' : '';
-    writeln("g.registerLazySingleton$asyncStr<${dep.type}>(()=> $constructor");
+
+    write("g.$funcName$asyncStr<${dep.type}>(()=> $constructor");
+
     closeRegisterFunc(dep);
     return buffer.toString();
   }