+++ injectable_generator/test/resolver/resolver_test.dart
@@ -1,0 +1,2 @@
+// Todo add resolver test
+void main() async {}